daily-input/v2functions/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
certifi==2022.5.18.1
charset-normalizer==2.0.12
greenlet==1.1.2
idna==3.3
numpy==1.21.6
panda==0.3.1
pandas==1.3.5
pyodbc==4.0.32
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.37
urllib3==1.26.9